refactor ram + storage
