+++ deploy-web/src/components/sdl/GpuFormControl.tsx
@@ -1,7 +1,7 @@
-import { Dispatch, ReactNode, SetStateAction } from "react";
+import { ReactNode } from "react";
 import { makeStyles } from "tss-react/mui";
 import { Box, Checkbox, CircularProgress, FormControl, FormHelperText, MenuItem, Select, Slider, TextField, Typography, useTheme } from "@mui/material";
-import { RentGpusFormValues, SdlBuilderFormValues, Service } from "@src/types";
+import { RentGpusFormValues, SdlBuilderFormValues } from "@src/types";
 import { CustomTooltip } from "../shared/CustomTooltip";
 import InfoIcon from "@mui/icons-material/Info";
 import { FormPaper } from "./FormPaper";
