+++ deploy-web/src/components/sdl/SimpleServiceFormControl.tsx
@@ -54,6 +54,8 @@ import { ExposeList } from "./ExposeList";
 import { PersistentStorage } from "./PersistentStorage";
 import { GpuFormControl } from "./GpuFormControl";
 import { CpuFormControl } from "./CpuFormControl";
+import { MemoryFormControl } from "./MemoryFormControl";
+import { StorageFormControl } from "./StorageFormControl";
 
 type Props = {
   _services: Service[];
@@ -325,236 +327,11 @@ export const SimpleServiceFormControl: React.FunctionComponent<Props> = ({
                 </Grid>
 
                 <Grid item xs={12}>
-                  <Controller
-                    control={control}
-                    name={`services.${serviceIndex}.profile.ram`}
-                    rules={{
-                      validate: v => {
-                        if (!v) return "Memory amount is required.";
-
-                        const currentUnit = memoryUnits.find(u => currentService.profile.ramUnit === u.suffix);
-                        const _value = (v || 0) * currentUnit.value;
-
-                        if (currentService.count === 1 && _value < minMemory) {
-                          return "Minimum amount of memory for a single service instance is 1 Mi.";
-                        } else if (currentService.count === 1 && currentService.count * _value > maxMemory) {
-                          return "Maximum amount of memory for a single service instance is 512 Gi.";
-                        } else if (currentService.count > 1 && currentService.count * _value > maxGroupMemory) {
-                          return "Maximum total amount of memory for a single service instance group is 1024 Gi.";
-                        }
-
-                        return true;
-                      }
-                    }}
-                    render={({ field, fieldState }) => (
-                      <FormPaper elevation={1} sx={{ padding: ".5rem 1rem", borderBottom: !!fieldState.error && `1px solid ${theme.palette.error.main}` }}>
-                        <FormControl
-                          className={cx(classes.formControl, classes.textField)}
-                          variant="standard"
-                          sx={{ marginBottom: "0 !important" }}
-                          error={!!fieldState.error}
-                        >
-                          <Box
-                            sx={{
-                              display: "flex",
-                              alignItems: { xs: "flex-start", sm: "center" },
-                              justifyContent: "space-between",
-                              flexDirection: { xs: "column", sm: "row" }
-                            }}
-                          >
-                            <Typography variant="body2" sx={{ display: "flex", alignItems: "center" }}>
-                              <MemoryIcon sx={{ color: theme.palette.grey[600], marginRight: ".5rem" }} fontSize="medium" />
-                              <strong>Memory</strong>
-
-                              <CustomTooltip
-                                arrow
-                                title={
-                                  <>
-                                    The amount of memory required for this workload.
-                                    <br />
-                                    <br />
-                                    The maximum for a single instance is 512 Gi.
-                                    <br />
-                                    <br />
-                                    The maximum total multiplied by the count of instances is 1024 Gi.
-                                  </>
-                                }
-                              >
-                                <InfoIcon color="disabled" fontSize="small" sx={{ marginLeft: "1rem" }} />
-                              </CustomTooltip>
-                            </Typography>
-
-                            <Box sx={{ marginTop: { xs: ".5rem", sm: 0 } }}>
-                              <TextField
-                                type="number"
-                                variant="outlined"
-                                error={!!fieldState.error}
-                                color="secondary"
-                                value={field.value || ""}
-                                onChange={event => field.onChange(parseFloat(event.target.value))}
-                                inputProps={{ min: 1, step: 1 }}
-                                size="small"
-                                sx={{ width: "100px" }}
-                              />
-
-                              <Controller
-                                control={control}
-                                name={`services.${serviceIndex}.profile.ramUnit`}
-                                rules={{ required: "Ram unit is required." }}
-                                defaultValue=""
-                                render={({ field }) => (
-                                  <Select
-                                    value={field.value || ""}
-                                    onChange={field.onChange}
-                                    variant="outlined"
-                                    size="small"
-                                    sx={{ width: "75px", marginLeft: ".25rem" }}
-                                    MenuProps={{ disableScrollLock: true }}
-                                  >
-                                    {memoryUnits.map(u => (
-                                      <MenuItem key={u.id} value={u.suffix}>
-                                        {u.suffix}
-                                      </MenuItem>
-                                    ))}
-                                  </Select>
-                                )}
-                              />
-                            </Box>
-                          </Box>
-
-                          <Slider
-                            value={field.value || 0}
-                            min={1}
-                            max={512}
-                            step={1}
-                            color="secondary"
-                            aria-label="RAM"
-                            valueLabelDisplay="auto"
-                            onChange={(event, newValue) => field.onChange(newValue)}
-                          />
-
-                          {!!fieldState.error && <FormHelperText>{fieldState.error.message}</FormHelperText>}
-                        </FormControl>
-                      </FormPaper>
-                    )}
-                  />
+                  <MemoryFormControl control={control as any} currentService={currentService} serviceIndex={serviceIndex} />
                 </Grid>
 
                 <Grid item xs={12}>
-                  <Controller
-                    control={control}
-                    rules={{
-                      validate: v => {
-                        if (!v) return "Storage amount is required.";
-
-                        const currentUnit = storageUnits.find(u => currentService.profile.storageUnit === u.suffix);
-                        const _value = (v || 0) * currentUnit.value;
-
-                        if (currentService.count * _value < minStorage) {
-                          return "Minimum amount of storage for a single service instance is 5 Mi.";
-                        } else if (currentService.count * _value > maxStorage) {
-                          return "Maximum amount of storage for a single service instance is 32 Ti.";
-                        }
-
-                        return true;
-                      }
-                    }}
-                    name={`services.${serviceIndex}.profile.storage`}
-                    render={({ field, fieldState }) => (
-                      <FormPaper elevation={1} sx={{ padding: ".5rem 1rem", borderBottom: !!fieldState.error && `1px solid ${theme.palette.error.main}` }}>
-                        <FormControl
-                          className={cx(classes.formControl, classes.textField)}
-                          variant="standard"
-                          sx={{ marginBottom: "0 !important" }}
-                          error={!!fieldState.error}
-                        >
-                          <Box
-                            sx={{
-                              display: "flex",
-                              alignItems: { xs: "flex-start", sm: "center" },
-                              justifyContent: "space-between",
-                              flexDirection: { xs: "column", sm: "row" }
-                            }}
-                          >
-                            <Typography variant="body2" sx={{ display: "flex", alignItems: "center" }}>
-                              <StorageIcon sx={{ color: theme.palette.grey[600], marginRight: ".5rem" }} fontSize="medium" />
-                              <strong>Ephemeral Storage</strong>
-
-                              <CustomTooltip
-                                arrow
-                                title={
-                                  <>
-                                    The amount of ephemeral disk storage required for this workload.
-                                    <br />
-                                    <br />
-                                    This disk storage is ephemeral, meaning it will be wiped out on every deployment update or provider reboot.
-                                    <br />
-                                    <br />
-                                    The maximum for a single instance is 32 Ti.
-                                    <br />
-                                    <br />
-                                    The maximum total multiplied by the count of instances is also 32 Ti.
-                                  </>
-                                }
-                              >
-                                <InfoIcon color="disabled" fontSize="small" sx={{ marginLeft: "1rem" }} />
-                              </CustomTooltip>
-                            </Typography>
-
-                            <Box sx={{ marginTop: { xs: ".5rem", sm: 0 } }}>
-                              <TextField
-                                type="number"
-                                variant="outlined"
-                                color="secondary"
-                                value={field.value || ""}
-                                error={!!fieldState.error}
-                                onChange={event => field.onChange(parseFloat(event.target.value))}
-                                inputProps={{ min: 1, step: 1 }}
-                                size="small"
-                                sx={{ width: "100px" }}
-                              />
-
-                              <Controller
-                                control={control}
-                                name={`services.${serviceIndex}.profile.storageUnit`}
-                                rules={{ required: "Storage unit is required." }}
-                                defaultValue=""
-                                render={({ field }) => (
-                                  <Select
-                                    value={field.value || ""}
-                                    onChange={field.onChange}
-                                    variant="outlined"
-                                    size="small"
-                                    sx={{ width: "75px", marginLeft: ".25rem" }}
-                                    MenuProps={{ disableScrollLock: true }}
-                                  >
-                                    {storageUnits.map(u => (
-                                      <MenuItem key={u.id} value={u.suffix}>
-                                        {u.suffix}
-                                      </MenuItem>
-                                    ))}
-                                  </Select>
-                                )}
-                              />
-                            </Box>
-                          </Box>
-
-                          <Slider
-                            value={field.value || 0}
-                            min={1}
-                            max={512}
-                            step={1}
-                            color="secondary"
-                            aria-label="Storage"
-                            valueLabelDisplay="auto"
-                            onChange={(event, newValue) => field.onChange(newValue)}
-                          />
-
-                          {!!fieldState.error && <FormHelperText>{fieldState.error.message}</FormHelperText>}
-                        </FormControl>
-                      </FormPaper>
-                    )}
-                  />
+                  <StorageFormControl control={control as any} currentService={currentService} serviceIndex={serviceIndex} />
                 </Grid>
 
                 <Grid item xs={12}>
